tags, and subset copying

+++ src/encord_active/lib/encord/project_sync.py
@@ -95,7 +95,9 @@ def _replace_uids(
     project_file_structure: ProjectFileStructure,
     renaming_map: dict[str, str],
 ):
-    original_mappings = json.loads(project_file_structure.mappings.read_text())
+    original_mappings = {}
+    if project_file_structure.mappings.is_file():
+        original_mappings = json.loads(project_file_structure.mappings.read_text())
 
     replace_in_files(project_file_structure, renaming_map)
     perform_db_fn_with_switched_paths(
@@ -105,7 +107,10 @@ def _replace_uids(
         update_embedding_identifiers(project_file_structure, embedding_type, renaming_map)
         update_2d_embedding_identifiers(project_file_structure, embedding_type, renaming_map)
 
-    new_mappings = {renaming_map[k]: v for k, v in original_mappings.items()}
+    if original_mappings:
+        new_mappings = {renaming_map[k]: v for k, v in original_mappings.items()}
+    else:
+        new_mappings = {v: k for k, v in renaming_map.items()}
     project_file_structure.mappings.write_text(json.dumps(new_mappings))
 
 
@@ -158,14 +163,16 @@ def copy_filtered_data(
 ):
     target_project_structure.data.mkdir(parents=True, exist_ok=True)
 
-    hash_mappings = json.loads(curr_project_structure.mappings.read_text())
-    filtered_hash_mappings = {
-        k: v for k, v in hash_mappings.items() if k in filtered_label_rows or k in filtered_data_hashes
-    }
-    target_project_structure.mappings.write_text(json.dumps(filtered_hash_mappings))
-    target_project_structure = ProjectFileStructure(
-        target_project_structure.project_dir
-    )  # Recreate object to reload mappings
+    filtered_hash_mappings = {}
+    if curr_project_structure.mappings.is_file():
+        hash_mappings = json.loads(curr_project_structure.mappings.read_text())
+        filtered_hash_mappings = {
+            k: v for k, v in hash_mappings.items() if k in filtered_label_rows or k in filtered_data_hashes
+        }
+        target_project_structure.mappings.write_text(json.dumps(filtered_hash_mappings))
+        target_project_structure = ProjectFileStructure(
+            target_project_structure.project_dir
+        )  # Recreate object to reload mappings
 
     for label_row_hash in filtered_label_rows:
         current_label_row_structure = curr_project_structure.label_row_structure(label_row_hash)
@@ -215,7 +222,8 @@ def copy_filtered_data(
 def create_filtered_db(target_project_dir: Path, filtered_df: pd.DataFrame):
     to_save_df = filtered_df.set_index("identifier")
     perform_db_fn_with_switched_paths(target_project_dir, lambda: MergedMetrics().replace_all(to_save_df))
-    perform_db_fn_with_switched_paths(target_project_dir, lambda: Tags().create_many(Tags().all()))
+    all_tags = Tags().all()
+    perform_db_fn_with_switched_paths(target_project_dir, lambda: Tags().create_many(all_tags))
 
 
 def perform_db_fn_with_switched_paths(target_project_dir: Path, f: Callable):
